Fix Link factory and proper test the presence of a boolean field.
According to the shoulda-matchers documentation:
We discourage using validate_inclusion_of with boolean columns.
In fact, there is never a case where a boolean column will be anything but true, false, or nil, as ActiveRecord will type-cast an incoming value to one of these three values.
That means there isn't any way we can refute this logic in a test.